Split files by browserify
