+++ app/assets/javascripts/serverspecs.js
@@ -5,14 +5,15 @@
 //
 // http://opensource.org/licenses/mit-license.php
 //
-//= require serverspec-gen/ast
-//= require serverspec-gen/ui
 
 //browserify functions for vue filters functionality
 var wrap = require('./modules/wrap');
 var listen = require('./modules/listen');
 var parseURLParams = require('./modules/getURL');
 var serverspecIndex = require('./modules/loadindex');
+
+require('serverspec-gen/ast');
+require('serverspec-gen/ui');
 
 var app;
