+++ frontend/src/serverspec-gen/ui.ts
@@ -1,6 +1,10 @@
 /// <reference path="../../declares.d.ts" />
 /// <reference path="./ast.ts" />
 
+import ResourcePanel from './resource-panel';
+import ItPanel       from './it-panel';
+
+
 class VueMain extends Vue {
   private ast: ASTInterfaces.Describe[];
   private rubyCode:     string;
@@ -35,9 +39,6 @@ class VueMain extends Vue {
   }
 
 
-
-
-
   // Computeds
 
   _rubyCode(): string {
@@ -49,75 +50,8 @@ class VueMain extends Vue {
     return `data:application/octet-stream,${encodeURIComponent(this.rubyCode)}`;
   }
 }
+const app = new VueMain([]);
 
-
-// TODO: Split files
-
-const ResourcePanel = Vue.extend({
-  template: '#resource-panel-template',
-  el: () => { return document.createElement('div'); },
-  props: {
-    desc: {
-      type: Object,
-      twoWay: true,
-      required: true,
-    },
-    idx: {
-      type: Number,
-      required: true,
-    },
-  },
-  methods: {
-    addIt: function () {
-      this.desc.body.push({
-        type: 'it',
-        should: true,
-        matcher: {name: 'be_', args: [], chains: []},
-      });
-    },
-
-    addIts: function () {
-      this.desc.body.push({
-        type: 'its',
-        name: "",
-        should: true,
-        matcher: {name: 'be_', args: [], chains: []},
-      });
-    },
-
-    removeIt: function(idx: number) {
-      console.log(idx);
-      this.desc.body.$remove(idx);
-    },
-  },
-  ready: function() {
-    console.log(this);
-  }
-});
 
 Vue.component("resource-panel", ResourcePanel);
-
-
-
-const ItPanel = Vue.extend({
-  template: '#it-panel-template',
-  el: () => {return document.createElement('div'); },
-  props: {
-    it: {
-      type: Object,
-      twoWay: true,
-      required: true,
-    },
-    idx: {
-      type: Number,
-      required: true,
-    }
-  },
-  ready: function() {
-    console.log(this);
-  },
-});
-
 Vue.component("it-panel", ItPanel);
-
-const app = new VueMain([]);
