+++ frontend/tsconfig.json
@@ -6,12 +6,17 @@
     "removeComments": true,
     "noLib": false,
     "rootDir": "src/",
-    "outDir": "dest/"
+    "outDir": "dest/",
+    "module": "commonjs"
   },
   "files": [
     "src/helper.ts",
     "src/server_status.ts",
     "src/users_admin.ts",
+    "src/serverspec-gen/ast.ts",
+    "src/serverspec-gen/it-panel.ts",
+    "src/serverspec-gen/resource-panel.ts",
+    "src/serverspec-gen/ui.ts",
     "src/models/base.ts",
     "src/models/cf_template.ts",
     "src/models/dish.ts",
@@ -21,8 +26,6 @@
     "src/models/rds_instance.ts",
     "src/models/resource.ts",
     "src/models/s3_bucket.ts",
-    "src/models/snapshot.ts",
-    "src/serverspec-gen/ast.ts",
-    "src/serverspec-gen/ui.ts"
+    "src/models/snapshot.ts"
   ]
 }
